python-api/requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
python-docx==1.2.0
sqlmodel==0.0.37
